+++ src/lib.rs
@@ -1,204 +1,11 @@
-//! # apalis-sqlite
-//!
-//! Background task processing for rust using apalis and sqlite.
-//!
-//! ## Features
-//!
-//! - **Reliable job queue** using SQLite as the backend.
-//! - **Multiple storage types**: standard polling and event-driven (hooked) storage.
-//! - **Custom codecs** for serializing/deserializing job arguments.
-//! - **Heartbeat and orphaned job re-enqueueing** for robust job processing.
-//! - **Integration with apalis workers and middleware.**
-//!
-//! ## Storage Types
-//!
-//! - [`SqliteStorage`]: Standard polling-based storage.
-//! - [`SqliteStorageWithHook`]: Event-driven storage using SQLite update hooks for low-latency job fetching.
-//! - [`SharedSqliteStorage`]: Shared storage for multiple job types.
-//!
-//! The naming is designed to clearly indicate the storage mechanism and its capabilities, but under the hood its the result is the `SqliteStorage` struct with different configurations.
-//!
-//! ## Examples
-//!
-//! ### Basic Worker Example
-//!
-//! ```rust
-//! # use apalis_sqlite::{SqliteStorage, SqlContext};
-//! # use apalis_core::task::Task;
-//! # use apalis_core::worker::context::WorkerContext;
-//! # use sqlx::SqlitePool;
-//! # use futures::stream;
-//! # use std::time::Duration;
-//! # use apalis_core::error::BoxDynError;
-//! # use futures::StreamExt;
-//! # use futures::SinkExt;
-//! # use apalis_core::worker::builder::WorkerBuilder;
-//! # use apalis_core::backend::TaskSink;
-//! #[tokio::main]
-//! async fn main() {
-//!     let pool = SqlitePool::connect(":memory:").await.unwrap();
-//!     SqliteStorage::setup(&pool).await.unwrap();
-//!     let mut backend = SqliteStorage::new(&pool);
-//!
-//!     let mut start = 0usize;
-//!     let mut items = stream::repeat_with(move || {
-//!         start += 1;
-//!         start
-//!     })
-//!     .take(10);
-//!     backend.push_stream(&mut items).await.unwrap();
-//!
-//!     async fn send_reminder(item: usize, wrk: WorkerContext) -> Result<(), BoxDynError> {
-//!         if item == 10 {
-//!             wrk.stop().unwrap();
-//!         }
-//!         Ok(())
-//!     }
-//!
-//!     let worker = WorkerBuilder::new("worker-1")
-//!         .backend(backend)
-//!         .build(send_reminder);
-//!     worker.run().await.unwrap();
-//! }
-//! ```
-//!
-//! ### Hooked Worker Example (Event-driven)
-//!
-//! ```rust,no_run
-//! # use apalis_sqlite::{SqliteStorage, SqlContext, Config};
-//! # use apalis_core::task::Task;
-//! # use apalis_core::worker::context::WorkerContext;
-//! # use apalis_core::backend::poll_strategy::{IntervalStrategy, StrategyBuilder};
-//! # use sqlx::SqlitePool;
-//! # use futures::stream;
-//! # use std::time::Duration;
-//! # use apalis_core::error::BoxDynError;
-//! # use futures::StreamExt;
-//! # use futures::SinkExt;
-//! # use apalis_core::worker::builder::WorkerBuilder;
-//!
-//! #[tokio::main]
-//! async fn main() {
-//!     let pool = SqlitePool::connect(":memory:").await.unwrap();
-//!     SqliteStorage::setup(&pool).await.unwrap();
-//!
-//!     let lazy_strategy = StrategyBuilder::new()
-//!         .apply(IntervalStrategy::new(Duration::from_secs(5)))
-//!         .build();
-//!     let config = Config::new("queue")
-//!         .with_poll_interval(lazy_strategy)
-//!         .set_buffer_size(5);
-//!     let backend = SqliteStorage::new_with_callback(&pool, &config);
-//!
-//!     tokio::spawn({
-//!         let pool = pool.clone();
-//!         let config = config.clone();
-//!         async move {
-//!             tokio::time::sleep(Duration::from_secs(2)).await;
-//!             let mut start = 0;
-//!             let items = stream::repeat_with(move || {
-//!                 start += 1;
-//!                 Task::builder(serde_json::to_vec(&start).unwrap())
-//!                     .run_after(Duration::from_secs(1))
-//!                     .with_ctx(SqlContext::new().with_priority(start))
-//!                     .build()
-//!             })
-//!             .take(20)
-//!             .collect::<Vec<_>>()
-//!             .await;
-//!             // push encoded tasks
-//!             apalis_sqlite::sink::push_tasks(pool, config, items).await.unwrap();
-//!         }
-//!     });
-//!
-//!     async fn send_reminder(item: usize, wrk: WorkerContext) -> Result<(), BoxDynError> {
-//!         if item == 1 {
-//!             apalis_core::timer::sleep(Duration::from_secs(1)).await;
-//!             wrk.stop().unwrap();
-//!         }
-//!         Ok(())
-//!     }
-//!
-//!     let worker = WorkerBuilder::new("worker-2")
-//!         .backend(backend)
-//!         .build(send_reminder);
-//!     worker.run().await.unwrap();
-//! }
-//! ```
-//! ### Workflow Example
-//!
-//! ```rust,no_run
-//! # use apalis_sqlite::{SqliteStorage, SqlContext, Config};
-//! # use apalis_core::task::Task;
-//! # use apalis_core::worker::context::WorkerContext;
-//! # use sqlx::SqlitePool;
-//! # use futures::stream;
-//! # use std::time::Duration;
-//! # use apalis_core::error::BoxDynError;
-//! # use futures::StreamExt;
-//! # use futures::SinkExt;
-//! # use apalis_core::worker::builder::WorkerBuilder;
-//! # use apalis_workflow::Workflow;
-//! # use apalis_workflow::WorkflowError;
-//! # use apalis_core::worker::event::Event;
-//! # use apalis_core::backend::WeakTaskSink;
-//! # use apalis_core::worker::ext::event_listener::EventListenerExt;
-//! #[tokio::main]
-//! async fn main() {
-//!     let workflow = Workflow::new("odd-numbers-workflow")
-//!         .then(|a: usize| async move {
-//!             Ok::<_, WorkflowError>((0..=a).collect::<Vec<_>>())
-//!         })
-//!         .filter_map(|x| async move {
-//!             if x % 2 != 0 { Some(x) } else { None }
-//!         })
-//!         .filter_map(|x| async move {
-//!             if x % 3 != 0 { Some(x) } else { None }
-//!         })
-//!         .filter_map(|x| async move {
-//!             if x % 5 != 0 { Some(x) } else { None }
-//!         })
-//!         .delay_for(Duration::from_millis(1000))
-//!         .then(|a: Vec<usize>| async move {
-//!             println!("Sum: {}", a.iter().sum::<usize>());
-//!             Ok::<(), WorkflowError>(())
-//!         });
-//!
-//!     let pool = SqlitePool::connect(":memory:").await.unwrap();
-//!     SqliteStorage::setup(&pool).await.unwrap();
-//!     let mut sqlite = SqliteStorage::new_in_queue(&pool, "test-workflow");
-//!
-//!     sqlite.push(100usize).await.unwrap();
-//!
-//!     let worker = WorkerBuilder::new("rango-tango")
-//!         .backend(sqlite)
-//!         .on_event(|ctx, ev| {
-//!             println!("On Event = {:?}", ev);
-//!             if matches!(ev, Event::Error(_)) {
-//!                 ctx.stop().unwrap();
-//!             }
-//!         })
-//!         .build(workflow);
-//!
-//!     worker.run().await.unwrap();
-//! }
-//! ```
-//!
-//! ## Observability
-//!
-//! You can track your jobs using [apalis-board](https://github.com/apalis-dev/apalis-board).
-//! ![Task](https://github.com/apalis-dev/apalis-board/raw/master/screenshots/task.png)
-//!
-//! ## License
-//!
-//! Licensed under either of Apache License, Version 2.0 or MIT license at your option.
+#![doc = include_str!("../README.md")]
 //!
 //! [`SqliteStorageWithHook`]: crate::SqliteStorage
 use std::{fmt, marker::PhantomData};
 
 use apalis_core::{
     backend::{
-        Backend, TaskStream,
+        Backend, BackendExt, TaskStream,
         codec::{Codec, json::JsonCodec},
     },
     features_table,
@@ -208,19 +15,21 @@ use apalis_core::{
 };
 pub use apalis_sql::context::SqlContext;
 use futures::{
-    FutureExt, StreamExt, TryFutureExt, TryStreamExt,
-    channel::mpsc,
+    FutureExt, Stream, StreamExt, TryFutureExt, TryStreamExt,
+    channel::mpsc::{self},
     future::ready,
     stream::{self, BoxStream, select},
 };
-use libsqlite3_sys::{sqlite3, sqlite3_update_hook};
-use sqlx::{Pool, Sqlite};
-use std::ffi::c_void;
+pub use sqlx::{
+    Connection, Pool, Sqlite, SqliteConnection, SqlitePool,
+    pool::{PoolConnection, PoolOptions},
+    sqlite::{SqliteConnectOptions, SqliteOperation},
+};
 use ulid::Ulid;
 
 use crate::{
     ack::{LockTaskLayer, SqliteAck},
-    callback::{HookCallbackListener, update_hook_callback},
+    callback::update_hook_callback,
     fetcher::{SqliteFetcher, SqlitePollFetcher, fetch_next},
     queries::{
         keep_alive::{initial_heartbeat, keep_alive, keep_alive_stream},
@@ -232,91 +41,24 @@ use crate::{
 mod ack;
 mod callback;
 mod config;
+/// Fetcher module for retrieving tasks from sqlite backend
 pub mod fetcher;
+mod from_row;
+/// Queries module for sqlite backend
 pub mod queries;
 mod shared;
+/// Sink module for pushing tasks to sqlite backend
 pub mod sink;
 
-mod from_row {
-    use chrono::{TimeZone, Utc};
-
-    #[derive(Debug)]
-    pub(crate) struct SqliteTaskRow {
-        pub(crate) job: Vec<u8>,
-        pub(crate) id: Option<String>,
-        pub(crate) job_type: Option<String>,
-        pub(crate) status: Option<String>,
-        pub(crate) attempts: Option<i64>,
-        pub(crate) max_attempts: Option<i64>,
-        pub(crate) run_at: Option<i64>,
-        pub(crate) last_result: Option<String>,
-        pub(crate) lock_at: Option<i64>,
-        pub(crate) lock_by: Option<String>,
-        pub(crate) done_at: Option<i64>,
-        pub(crate) priority: Option<i64>,
-        pub(crate) metadata: Option<String>,
-    }
-
-    impl TryInto<apalis_sql::from_row::TaskRow> for SqliteTaskRow {
-        type Error = sqlx::Error;
-
-        fn try_into(self) -> Result<apalis_sql::from_row::TaskRow, Self::Error> {
-            Ok(apalis_sql::from_row::TaskRow {
-                job: self.job,
-                id: self
-                    .id
-                    .ok_or_else(|| sqlx::Error::Protocol("Missing id".into()))?,
-                job_type: self
-                    .job_type
-                    .ok_or_else(|| sqlx::Error::Protocol("Missing job_type".into()))?,
-                status: self
-                    .status
-                    .ok_or_else(|| sqlx::Error::Protocol("Missing status".into()))?,
-                attempts: self
-                    .attempts
-                    .ok_or_else(|| sqlx::Error::Protocol("Missing attempts".into()))?
-                    as usize,
-                max_attempts: self.max_attempts.map(|v| v as usize),
-                run_at: self.run_at.map(|ts| {
-                    Utc.timestamp_opt(ts, 0)
-                        .single()
-                        .ok_or_else(|| sqlx::Error::Protocol("Invalid run_at timestamp".into()))
-                        .unwrap()
-                }),
-                last_result: self
-                    .last_result
-                    .map(|res| serde_json::from_str(&res).unwrap_or(serde_json::Value::Null)),
-                lock_at: self.lock_at.map(|ts| {
-                    Utc.timestamp_opt(ts, 0)
-                        .single()
-                        .ok_or_else(|| sqlx::Error::Protocol("Invalid run_at timestamp".into()))
-                        .unwrap()
-                }),
-                lock_by: self.lock_by,
-                done_at: self.done_at.map(|ts| {
-                    Utc.timestamp_opt(ts, 0)
-                        .single()
-                        .ok_or_else(|| sqlx::Error::Protocol("Invalid run_at timestamp".into()))
-                        .unwrap()
-                }),
-                priority: self.priority.map(|v| v as usize),
-                metadata: self
-                    .metadata
-                    .map(|meta| serde_json::from_str(&meta).unwrap_or(serde_json::Value::Null)),
-            })
-        }
-    }
-}
-
+/// Type alias for a task stored in sqlite backend
 pub type SqliteTask<Args> = Task<Args, SqlContext, Ulid>;
-pub use callback::{CallbackListener, DbEvent};
+pub use callback::{DbEvent, HookCallbackListener};
 pub use config::Config;
-pub use shared::{SharedPostgresError, SharedSqliteStorage};
-pub use sqlx::SqlitePool;
+pub use shared::{SharedSqliteError, SharedSqliteStorage};
 
+/// CompactType is the type used for compact serialization in sqlite backend
 pub type CompactType = Vec<u8>;
 
-const INSERT_OPERATION: &str = "INSERT";
 const JOBS_TABLE: &str = "Jobs";
 
 /// SqliteStorage is a storage backend for apalis using sqlite as the database.
@@ -337,11 +79,11 @@ const JOBS_TABLE: &str = "Jobs";
     Backend => supported("Supports storage and retrieval of tasks", true),
     TaskSink => supported("Ability to push new tasks", true),
     Serialization => supported("Serialization support for arguments", true),
+    Workflow => supported("Flexible enough to support workflows", true),
     WebUI => supported("Expose a web interface for monitoring tasks", true),
     FetchById => supported("Allow fetching a task by its ID", false),
     RegisterWorker => supported("Allow registering a worker with the backend", false),
     MakeShared => supported("Share one connection across multiple workers via [`SharedSqliteStorage`]", false),
-    Workflow => supported("Flexible enough to support workflows", true),
     WaitForCompletion => supported("Wait for tasks to complete without blocking", true),
     ResumeById => supported("Resume a task by its ID", false),
     ResumeAbandoned => supported("Resume abandoned tasks", false),
@@ -373,7 +115,7 @@ impl<T, C, F> fmt::Debug for SqliteStorage<T, C, F> {
 
 impl<T, C, F: Clone> Clone for SqliteStorage<T, C, F> {
     fn clone(&self) -> Self {
-        SqliteStorage {
+        Self {
             sink: self.sink.clone(),
             pool: self.pool.clone(),
             job_type: PhantomData,
@@ -404,6 +146,7 @@ impl SqliteStorage<(), (), ()> {
 
     /// Get sqlite migrations without running them
     #[cfg(feature = "migrate")]
+    #[must_use]
     pub fn migrations() -> sqlx::migrate::Migrator {
         sqlx::migrate!("./migrations")
     }
@@ -411,13 +154,10 @@ impl SqliteStorage<(), (), ()> {
 
 impl<T> SqliteStorage<T, (), ()> {
     /// Create a new SqliteStorage
+    #[must_use]
     pub fn new(
         pool: &Pool<Sqlite>,
-    ) -> SqliteStorage<
-        T,
-        JsonCodec<CompactType>,
-        fetcher::SqliteFetcher<T, CompactType, JsonCodec<CompactType>>,
-    > {
+    ) -> SqliteStorage<T, JsonCodec<CompactType>, fetcher::SqliteFetcher> {
         let config = Config::new(std::any::type_name::<T>());
         SqliteStorage {
             pool: pool.clone(),
@@ -425,20 +165,16 @@ impl<T> SqliteStorage<T, (), ()> {
             sink: SqliteSink::new(pool, &config),
             config,
             codec: PhantomData,
-            fetcher: fetcher::SqliteFetcher {
-                _marker: PhantomData,
-            },
+            fetcher: fetcher::SqliteFetcher,
         }
     }
 
+    /// Create a new SqliteStorage for a specific queue
+    #[must_use]
     pub fn new_in_queue(
         pool: &Pool<Sqlite>,
         queue: &str,
-    ) -> SqliteStorage<
-        T,
-        JsonCodec<CompactType>,
-        fetcher::SqliteFetcher<T, CompactType, JsonCodec<CompactType>>,
-    > {
+    ) -> SqliteStorage<T, JsonCodec<CompactType>, fetcher::SqliteFetcher> {
         let config = Config::new(queue);
         SqliteStorage {
             pool: pool.clone(),
@@ -446,84 +182,82 @@ impl<T> SqliteStorage<T, (), ()> {
             sink: SqliteSink::new(pool, &config),
             config,
             codec: PhantomData,
-            fetcher: fetcher::SqliteFetcher {
-                _marker: PhantomData,
-            },
+            fetcher: fetcher::SqliteFetcher,
         }
     }
 
-    pub fn new_with_codec<Codec>(
-        pool: &Pool<Sqlite>,
-        config: &Config,
-    ) -> SqliteStorage<T, Codec, fetcher::SqliteFetcher<T, CompactType, Codec>> {
-        SqliteStorage {
-            pool: pool.clone(),
-            job_type: PhantomData,
-            config: config.clone(),
-            codec: PhantomData,
-            sink: SqliteSink::new(pool, config),
-            fetcher: fetcher::SqliteFetcher {
-                _marker: PhantomData,
-            },
-        }
-    }
-
+    /// Create a new SqliteStorage with config
+    #[must_use]
     pub fn new_with_config(
         pool: &Pool<Sqlite>,
         config: &Config,
-    ) -> SqliteStorage<
-        T,
-        JsonCodec<CompactType>,
-        fetcher::SqliteFetcher<T, CompactType, JsonCodec<CompactType>>,
-    > {
+    ) -> SqliteStorage<T, JsonCodec<CompactType>, fetcher::SqliteFetcher> {
         SqliteStorage {
             pool: pool.clone(),
             job_type: PhantomData,
             config: config.clone(),
             codec: PhantomData,
             sink: SqliteSink::new(pool, config),
-            fetcher: fetcher::SqliteFetcher {
-                _marker: PhantomData,
-            },
+            fetcher: fetcher::SqliteFetcher,
         }
     }
 
+    /// Create a new SqliteStorage with hook callback listener
+    #[must_use]
     pub fn new_with_callback(
-        pool: &Pool<Sqlite>,
+        url: &str,
         config: &Config,
     ) -> SqliteStorage<T, JsonCodec<CompactType>, HookCallbackListener> {
-        SqliteStorage {
-            pool: pool.clone(),
-            job_type: PhantomData,
-            config: config.clone(),
-            codec: PhantomData,
-            sink: SqliteSink::new(pool, config),
-            fetcher: HookCallbackListener,
-        }
-    }
+        let (tx, rx) = mpsc::unbounded::<DbEvent>();
 
-    pub fn new_with_codec_callback<Codec>(
-        pool: &Pool<Sqlite>,
-        config: &Config,
-    ) -> SqliteStorage<T, Codec, HookCallbackListener> {
+        let listener = HookCallbackListener::new(rx);
+        let pool = PoolOptions::<Sqlite>::new()
+            .after_connect(move |conn, _meta| {
+                let mut tx = tx.clone();
+                Box::pin(async move {
+                    let mut lock_handle = conn.lock_handle().await?;
+                    lock_handle.set_update_hook(move |ev| update_hook_callback(ev, &mut tx));
+                    Ok(())
+                })
+            })
+            .connect_lazy(url)
+            .expect("Failed to create Sqlite pool");
         SqliteStorage {
             pool: pool.clone(),
             job_type: PhantomData,
             config: config.clone(),
             codec: PhantomData,
-            sink: SqliteSink::new(pool, config),
-            fetcher: HookCallbackListener,
+            sink: SqliteSink::new(&pool, config),
+            fetcher: listener,
         }
     }
 }
 
 impl<T, C, F> SqliteStorage<T, C, F> {
+    /// Change the codec used for serialization/deserialization
+    pub fn with_codec<D>(self) -> SqliteStorage<T, D, F> {
+        SqliteStorage {
+            sink: SqliteSink::new(&self.pool, &self.config),
+            pool: self.pool,
+            job_type: PhantomData,
+            config: self.config,
+            codec: PhantomData,
+            fetcher: self.fetcher,
+        }
+    }
+
+    /// Get the config used by the storage
     pub fn config(&self) -> &Config {
         &self.config
     }
+
+    /// Get the connection pool used by the storage
+    pub fn pool(&self) -> &Pool<Sqlite> {
+        &self.pool
+    }
 }
 
-impl<Args, Decode> Backend for SqliteStorage<Args, Decode, SqliteFetcher<Args, CompactType, Decode>>
+impl<Args, Decode> Backend for SqliteStorage<Args, Decode, SqliteFetcher>
 where
     Args: Send + 'static + Unpin,
     Decode: Codec<Args, Compact = CompactType> + 'static + Send,
@@ -533,10 +267,6 @@ where
     type IdType = Ulid;
 
     type Context = SqlContext;
-
-    type Codec = Decode;
-
-    type Compact = CompactType;
 
     type Error = sqlx::Error;
 
@@ -567,20 +297,32 @@ where
     }
 
     fn poll(self, worker: &WorkerContext) -> Self::Stream {
-        let fut = initial_heartbeat(
-            self.pool.clone(),
-            self.config().clone(),
-            worker.clone(),
-            "SqliteStorage",
-        );
-        let register = stream::once(fut.map(|_| Ok(None)));
-        register
-            .chain(SqlitePollFetcher::<Args, CompactType, Decode>::new(
-                &self.pool,
-                &self.config,
-                worker,
-            ))
+        self.poll_default(worker)
+            .map(|a| match a {
+                Ok(Some(task)) => Ok(Some(
+                    task.try_map(|t| Decode::decode(&t))
+                        .map_err(|e| sqlx::Error::Decode(e.into()))?,
+                )),
+                Ok(None) => Ok(None),
+                Err(e) => Err(e),
+            })
             .boxed()
+    }
+}
+
+impl<Args, Decode: Send + 'static> BackendExt for SqliteStorage<Args, Decode, SqliteFetcher>
+where
+    Self: Backend<Args = Args, IdType = Ulid, Context = SqlContext, Error = sqlx::Error>,
+    Decode: Codec<Args, Compact = CompactType> + Send + 'static,
+    Decode::Error: std::error::Error + Send + Sync + 'static,
+    Args: Send + 'static + Unpin,
+{
+    type Codec = Decode;
+    type Compact = CompactType;
+    type CompactStream = TaskStream<SqliteTask<Self::Compact>, sqlx::Error>;
+
+    fn poll_compact(self, worker: &WorkerContext) -> Self::CompactStream {
+        self.poll_default(worker).boxed()
     }
 }
 
@@ -595,10 +337,6 @@ where
 
     type Context = SqlContext;
 
-    type Codec = Decode;
-
-    type Compact = CompactType;
-
     type Error = sqlx::Error;
 
     type Stream = TaskStream<SqliteTask<Args>, sqlx::Error>;
@@ -628,10 +366,41 @@ where
     }
 
     fn poll(self, worker: &WorkerContext) -> Self::Stream {
-        let (tx, rx) = mpsc::unbounded::<DbEvent>();
+        self.poll_with_listener(worker)
+            .map(|a| match a {
+                Ok(Some(task)) => Ok(Some(
+                    task.try_map(|t| Decode::decode(&t))
+                        .map_err(|e| sqlx::Error::Decode(e.into()))?,
+                )),
+                Ok(None) => Ok(None),
+                Err(e) => Err(e),
+            })
+            .boxed()
+    }
+}
 
-        let listener = CallbackListener::new(rx);
+impl<Args, Decode: Send + 'static> BackendExt for SqliteStorage<Args, Decode, HookCallbackListener>
+where
+    Self: Backend<Args = Args, IdType = Ulid, Context = SqlContext, Error = sqlx::Error>,
+    Decode: Codec<Args, Compact = CompactType> + Send + 'static,
+    Decode::Error: std::error::Error + Send + Sync + 'static,
+    Args: Send + 'static + Unpin,
+{
+    type Codec = Decode;
+    type Compact = CompactType;
+    type CompactStream = TaskStream<SqliteTask<Self::Compact>, sqlx::Error>;
 
+    fn poll_compact(self, worker: &WorkerContext) -> Self::CompactStream {
+        self.poll_with_listener(worker).boxed()
+    }
+}
+
+impl<Args, Decode: Send + 'static> SqliteStorage<Args, Decode, HookCallbackListener> {
+    fn poll_with_listener(
+        self,
+        worker: &WorkerContext,
+    ) -> impl Stream<Item = Result<Option<SqliteTask<CompactType>>, sqlx::Error>> + Send + 'static
+    {
         let pool = self.pool.clone();
         let config = self.config.clone();
         let worker = worker.clone();
@@ -641,31 +410,19 @@ where
             worker.clone(),
             "SqliteStorageWithHook",
         );
-        let p = pool.clone();
         let register_worker = stream::once(register_worker.map_ok(|_| None));
-        let register_listener = stream::once(async move {
-            // This is still a little tbd, but the idea is to test the update hook
-            let mut conn = p.acquire().await?;
-            // Get raw sqlite3* handle
-            let handle: *mut sqlite3 = conn.lock_handle().await.unwrap().as_raw_handle().as_ptr();
-
-            // Put sender in a Box so it has a stable memory address
-            let tx_box = Box::new(tx);
-            let tx_ptr = Box::into_raw(tx_box) as *mut c_void;
-            unsafe {
-                sqlite3_update_hook(handle, Some(update_hook_callback), tx_ptr);
-            }
-            Ok(None)
-        });
-        let eager_fetcher: SqlitePollFetcher<Args, CompactType, Decode> =
+        let eager_fetcher: SqlitePollFetcher<CompactType, Decode> =
             SqlitePollFetcher::new(&self.pool, &self.config, &worker);
-        let lazy_fetcher = listener
-            .filter(|a| ready(a.operation() == INSERT_OPERATION && a.table_name() == JOBS_TABLE))
+        let lazy_fetcher = self
+            .fetcher
+            .filter(|a| {
+                ready(a.operation() == &SqliteOperation::Insert && a.table_name() == JOBS_TABLE)
+            })
             .inspect(|db_event| {
-                log::debug!("Received DB event: {db_event:?}");
+                log::debug!("Received new job event: {db_event:?}");
             })
             .ready_chunks(self.config.buffer_size())
-            .then(move |_| fetch_next::<Args, Decode>(pool.clone(), config.clone(), worker.clone()))
+            .then(move |_| fetch_next(pool.clone(), config.clone(), worker.clone()))
             .flat_map(|res| match res {
                 Ok(tasks) => stream::iter(tasks).map(Ok).boxed(),
                 Err(e) => stream::iter(vec![Err(e)]).boxed(),
@@ -675,10 +432,28 @@ where
                 Err(e) => Err(e),
             });
 
-        register_worker
-            .chain(register_listener)
-            .chain(select(lazy_fetcher, eager_fetcher))
-            .boxed()
+        register_worker.chain(select(lazy_fetcher, eager_fetcher))
+    }
+}
+
+impl<Args, Decode: Send + 'static, F> SqliteStorage<Args, Decode, F> {
+    fn poll_default(
+        self,
+        worker: &WorkerContext,
+    ) -> impl Stream<Item = Result<Option<SqliteTask<CompactType>>, sqlx::Error>> + Send + 'static
+    {
+        let fut = initial_heartbeat(
+            self.pool.clone(),
+            self.config().clone(),
+            worker.clone(),
+            "SqliteStorage",
+        );
+        let register = stream::once(fut.map(|_| Ok(None)));
+        register.chain(SqlitePollFetcher::<CompactType, Decode>::new(
+            &self.pool,
+            &self.config,
+            worker,
+        ))
     }
 }
 
@@ -686,19 +461,11 @@ where
 mod tests {
     use std::time::Duration;
 
-    use apalis_workflow::{Workflow, WorkflowError};
+    use apalis::prelude::*;
+    use apalis_workflow::*;
     use chrono::Local;
-
-    use apalis_core::{
-        backend::{
-            WeakTaskSink,
-            poll_strategy::{IntervalStrategy, StrategyBuilder},
-        },
-        error::BoxDynError,
-        task::data::Data,
-        worker::{builder::WorkerBuilder, event::Event, ext::event_listener::EventListenerExt},
-    };
     use serde::{Deserialize, Serialize};
+    use sqlx::SqlitePool;
 
     use super::*;
 
@@ -737,8 +504,6 @@ mod tests {
     #[tokio::test]
     async fn hooked_worker() {
         const ITEMS: usize = 20;
-        let pool = SqlitePool::connect(":memory:").await.unwrap();
-        SqliteStorage::setup(&pool).await.unwrap();
 
         let lazy_strategy = StrategyBuilder::new()
             .apply(IntervalStrategy::new(Duration::from_secs(5)))
@@ -746,7 +511,9 @@ mod tests {
         let config = Config::new("rango-tango-queue")
             .with_poll_interval(lazy_strategy)
             .set_buffer_size(5);
-        let backend = SqliteStorage::new_with_callback(&pool, &config);
+        let backend = SqliteStorage::new_with_callback(":memory:", &config);
+        let pool = backend.pool().clone();
+        SqliteStorage::setup(&pool).await.unwrap();
 
         tokio::spawn(async move {
             tokio::time::sleep(Duration::from_secs(2)).await;
@@ -756,14 +523,15 @@ mod tests {
                 start += 1;
 
                 Task::builder(serde_json::to_vec(&start).unwrap())
-                    // .run_after(Duration::from_secs(1))
                     .with_ctx(SqlContext::new().with_priority(start))
                     .build()
             })
             .take(ITEMS)
             .collect::<Vec<_>>()
             .await;
-            sink::push_tasks(pool, config, items).await.unwrap();
+            crate::sink::push_tasks(pool.clone(), config, items)
+                .await
+                .unwrap();
         });
 
         async fn send_reminder(item: usize, wrk: WorkerContext) -> Result<(), BoxDynError> {
@@ -784,20 +552,18 @@ mod tests {
     #[tokio::test]
     async fn test_workflow() {
         let workflow = Workflow::new("odd-numbers-workflow")
-            .then(|a: usize| async move { Ok::<_, WorkflowError>((0..=a).collect::<Vec<_>>()) })
+            .and_then(|a: usize| async move { Ok::<_, BoxDynError>((0..=a).collect::<Vec<_>>()) })
             .filter_map(|x| async move { if x % 2 != 0 { Some(x) } else { None } })
             .filter_map(|x| async move { if x % 3 != 0 { Some(x) } else { None } })
             .filter_map(|x| async move { if x % 5 != 0 { Some(x) } else { None } })
             .delay_for(Duration::from_millis(1000))
-            .then(|a: Vec<usize>| async move {
+            .and_then(|a: Vec<usize>| async move {
                 println!("Sum: {}", a.iter().sum::<usize>());
-                Err::<(), WorkflowError>(WorkflowError::MissingContextError)
+                Err::<(), BoxDynError>("Intentional Error".into())
             });
 
-        let pool = SqlitePool::connect(":memory:").await.unwrap();
-        SqliteStorage::setup(&pool).await.unwrap();
         let mut sqlite = SqliteStorage::new_with_callback(
-            &pool,
+            ":memory:",
             &Config::new("workflow-queue").with_poll_interval(
                 StrategyBuilder::new()
                     .apply(IntervalStrategy::new(Duration::from_millis(100)))
@@ -805,7 +571,9 @@ mod tests {
             ),
         );
 
-        sqlite.push(100usize).await.unwrap();
+        SqliteStorage::setup(sqlite.pool()).await.unwrap();
+
+        sqlite.push_start(100usize).await.unwrap();
 
         let worker = WorkerBuilder::new("rango-tango")
             .backend(sqlite)
@@ -847,16 +615,16 @@ mod tests {
 
         let workflow = Workflow::new("text-pipeline")
             // Step 1: Preprocess input (e.g., tokenize, lowercase)
-            .then(|input: UserInput, mut worker: WorkerContext| async move {
-                worker.emit(&Event::Custom(Box::new(format!(
+            .and_then(|input: UserInput, mut worker: WorkerContext| async move {
+                worker.emit(&Event::custom(format!(
                     "Preprocessing input: {}",
                     input.text
-                ))));
+                )));
                 let processed = input.text.to_lowercase();
-                Ok::<_, WorkflowError>(processed)
+                Ok::<_, BoxDynError>(processed)
             })
             // Step 2: Classify text
-            .then(|text: String| async move {
+            .and_then(|text: String| async move {
                 let confidence = 0.85; // pretend model confidence
                 let items = text.split_whitespace().collect::<Vec<_>>();
                 let results = items
@@ -872,7 +640,7 @@ mod tests {
                         confidence,
                     })
                     .collect::<Vec<_>>();
-                Ok::<_, WorkflowError>(results)
+                Ok::<_, BoxDynError>(results)
             })
             // Step 3: Filter out low-confidence predictions
             .filter_map(
@@ -900,8 +668,7 @@ mod tests {
                     })
                 }
             })
-            .then(|a: Vec<Summary>, mut worker: WorkerContext| async move {
-                dbg!(&a);
+            .and_then(|a: Vec<Summary>, mut worker: WorkerContext| async move {
                 worker.emit(&Event::Custom(Box::new(format!(
                     "Generated {} summaries",
                     a.len()
@@ -909,14 +676,15 @@ mod tests {
                 worker.stop()
             });
 
-        let pool = SqlitePool::connect(":memory:").await.unwrap();
-        SqliteStorage::setup(&pool).await.unwrap();
-        let mut sqlite = SqliteStorage::new_with_callback(&pool, &Config::new("text-pipeline"));
+        let mut sqlite =
+            SqliteStorage::new_with_callback(":memory:", &Config::new("text-pipeline"));
+
+        SqliteStorage::setup(sqlite.pool()).await.unwrap();
 
         let input = UserInput {
             text: "Rust makes systems programming delightful!".to_string(),
         };
-        sqlite.push(input).await.unwrap();
+        sqlite.push_start(input).await.unwrap();
 
         let worker = WorkerBuilder::new("rango-tango")
             .backend(sqlite)